Do not check euid in buxton_direct_open
The symbol is only usable from libbuxton and is not publicly exported,
therefore only programs and libraries built in tree can access this method.
Also, we'll eventually be running as the 'buxton' user which will not have
euid of 0.

Also fixed a small issue whereby _directPermitted may still not be
initialised when buxton_client_get_value has been called, which is
an error.