Start to implement configuration loading
This commit adds the skeleton of our configuration loading. We essentially
check all files in CONFIG_DIRECTORY for the *.layer extension. If we find
them, we try to load using iniparser and populate a dictionary *.

Following commits will flesh out this code and implement the actual
configuration (layer names, smack labels and database modules).

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>