Implement --direct option in buxtonctl (cli)
The root user may use the --direct option to directly manipulate
values in buxton layers in the very same way that the buxton daemon
will.

This must be the very first option to buxtonctl, and the user must
have the euid of 0 to use this. For example:

   buxtonctl --direct set-string resource key value

In the future, it will only be possible to use the --direct option
from within packaging scripts from vendor and when the daemon is
offline.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>