Add utility method to form database path
Each of the database modules will need to be able to determine
the path of the database to open/create.

It should be noted that when using user layers, the uid should
already of the layer should already be set. This is relatively
simple as the next commit will switch the backends to accept
BuxtonLayer instead of char* references.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>