Implement get_value in gdbm backend and lbuxton
Quite simply we can now retrieve an entire BuxtonData struct from
the gdbm backend now. In order to avoid internal problems we create
a deep copy of the struct so that we do not return the internal pointer.

Client user should free data.store.d_string, as this is set using strdup.
Example use of get:

	BuxtonData get;
	buxton_client_get_value(&client, "base", "KeyName", &get);

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>