Friendler CLI fixes
This commit makes the CLI a little friendlier by informing the user they've
requested one type of key, and giving the string representation of the key
they really wanted.

We could of course make get_value handle all types without needing a requested
type, but this is not good for anyone interfacing with buxton via the CLI.
With these changes the user is forced to use the correct type, so that they're
aware of its type before modifying it.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>