Convert pollfds to use greedy_realloc()
This allows bt-daemon to accept more than just 32 clients total
in the future, without it, we'll hit a problem once we start
doing tests with many clients (performance).

Copied greedy_realloc() from systemd. It seems to work, but real
testing still needs to be done to assure everything works OK.