libbuxton: Only use direct when BUXTON_SOCKET doesn't exist
In order to prevent database files from being attempted to be used by
the buxton daemon and the client directly, make direct check if the
BUXTON_SOCKET file exists first. If it does exist just use the non
direct open as this works with our existing use cases of testing and
package installation.