libbuxton: Add function to return data based on priority
While it is still useful especially for testing to get data based on
key and layer, the main purpose of buxton's layers is to get the
correct value based on priority.

So iterate through all the layers to find which ones contain the key
being sought and keep track of which layer that had the key, had the
highest priority and return that one's data.