daemon: Send notification upon key change
As we rely on active connections and credentials may change, we only
inform the interested client that the key changed, and which key
it is.

We do not inform the client of the layer on which the key changed,
nor do we send them the value. An interested client should then
ask Buxton for the value of the new key using their active client
connection.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>