shared: Correct lists being used in notifications
Notification functions should be iterating over notification lists
instead of client lists (and should just be referencing the client in
the notification list item). We'll need to add ways to cleanup defunct
clients from the notification list.

Probably best to have an added feature in the poll loop that will have
a list of keys that have been updated or a boolean that a key was set
so we'll know to purge our defunct client list and then send out
notifications.

Signed-off-by: William Douglas <william.douglas@intel.com>