Convert daemon, protocol and serialize to use BuxtonArray
We can now begin the process of cleaning up the code, by using bool's
instead of relying on size_t returns being larger than 0, as the
length is always recorded in a BuxtonArray.

We'll also need to add utility functions to free BuxtonData automatically
in the future via buxton_array_free, to reduce any duplication.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>