shared: Set BuxtonData's value's length correctly based on type
Because the BuxtonDataStore's BuxtonString element has a char * in it
the size of the store would vary between 32 and 64bit architectures.
In order to allow sending the correct size in an architecture agnostic
way, set p_length based on the BuxtonData type instead of based on the
union max size (could be 16 or 8 bytes depending on architecture).