Begin migration to BuxtonControl
Currently direct and wire-based clients share the same API, potentially
enabling non-direct clients to use non-client code.

We will begin a migration to BuxtonControl, which can only be used by
the CLI tool with --direct mode, and by bt-daemon. Direct users are
to become responsible for the state and internal tracking of buxton,
such as maintaining loaded backends and layers, as opposed to retaining
these as global variables within libbuxton.

By removing this functionality, we can drop many redundant checks,
and trim the public section of the library further. Currently the
BuxtonConfig struct is a placeholder, which will be utilised when
backend implementation specifics are migrated from libbuxton to
backend.{c,h}

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>