serialize: Support serialization of BuxtonArray
With this commit we can now pass a BuxtonArray to buxton_serialize_message
as the only parameter, using BUXTON_CONTROL_LIST, to serialize all of
the BuxtonData's within that array.

Deserialization is unaffected, as the returned data is correctly
serialized with the parameter count coming from the length of the
original array.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>