Implement wire-protocol side of list_keys
We merely return a BuxtonArray for the client to transparently deal
with, populated with the key names found in a given layer. A couple
of other changes needed making, such as the incorrect null-char
detection in the deserialisation code.

Given that the list method still needs to send a status, and a
complete BuxtonArray as a parameter, the serialize function now
handles this as a special case, ensuring that the message type
is sent first (BUXTON_STATUS)

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>