daemon: fix client Smack label handling
The fgetxattr() call here is for the client sockfd on the server, so the
Smack label returned is not the one we need. In fact, it label value
appears to be floor regardless of the label bt-daemon runs with.

To fix the issue, we need the actual label from the client end of the
socket, found in the SO_PEERSEC option value for the client sockfd.