Switch to configurator
Configurator is great.  It replaces the usage of constants.[ch] to a
flexible system.  It also encapsulates the ini parser functionality to
one spot.  The tests are changed to more sensible directories.

Configurator introduces order of precedence for configuration items.
1) things added through the command line with buxton_add_cmd_line()
2) things in the environment
3) things in the conf file
4) compiled in defaults

Configurator also adds a section to the configuration file to store
these values.

Configurator improves your day and brightens your future

Signed-off-by: Michael Leibowitz <michael.leibowitz@intel.com>