core: Fix possible starvation when handling multiple clients
Instead of reading from a client fd until there is no longer data,
have a max number of messages that the daemon processes in one poll.
Then return with a status indicating more messages are available to be
processed and set the poll to timeout to 0 in order to quickly return
to processing new data from clients.