Remove assert that breaks things
The assertion was misplaced, because it should not be fatal in some
circumstances to fail to load the conf file.  For example, in the client
library, having the socket specified either via env or command line
should be the only requirement.  This removes the assert, which in
addition to being wrong also made tests failed with asserts are
enabled.  The author regrets the error.

Signed-off-by: Michael Leibowitz <michael.leibowitz@intel.com>