db/gdbm: Store deletion statistics
While storing keys in the database automatically updates the file
structure, deletions (unset) can result in sparse/unpopulated
areas in the file, slowing I/O and a large file size in general.

We store the deletion stats so that when they reach a peak
size they can be reorganized (TODO).

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>