Work-around a libtool nanny feature
Libtool is "smart."  Libtool will use -Wl,whole-archive on convenience
libraries, which it automatically detects in conjunction with its evil
paternalistic twin, automake, through noinst_LTLIBRARIES.  Of course,
this works with linking your convenience library to shared objects.  But
libtool has deemed it improper to link convenience libraries into your
programs.  They have the following to say:

   As a rule of thumb, link a libtool convenience library into at most
   one libtool library, and never into a program, and link libtool
   static convenience libraries only into programs, and only if you need
   to carry library dependency information to the user of the static
   convenience library.

What this does not really tell you is that libtool doesn't use the
-Wl,whole-archive flags when linking the library into your program, only
your shared libraries.  So, your program will go on its merry way
without the symbols you so desperately need.  That is, until it dlopen's
a plugin that uses the symbols that aren't there.  Then it's "unresolved
symbol!", for you.  At this point, your program usually bites the wax
tadpole.  As such, we have included useless and ugly functions into
buxtonctl to make sure the symbols in the convenience library are there
for the direct mode of operation, where the plugins are loaded by
buxtonctl.

tl;dr my code is ugly- it's libtool's fault

Signed-off-by: Michael Leibowitz <michael.leibowitz@intel.com>