Remove iniparser from Buxton
We no longer need to ship an internal copy of iniparser as it
is available in the software repositories of several distributions
already.

Sometimes iniparser is shipped without a .pc file, meaning we
need to manually check for the existence of the headers and
library, ensuring they'll work for Buxton.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>