Add a new Hashmap implementation
BuxtonHashmap is a simple integer-key based hashmap with a focus
on speed and performance. It is intended to replace the use of
client_list_item (LIST) and enable rapid lookups via file descriptor,
or other key, to the connected client list.

It works very similarly to other hashmap implementations, except
it merely stores pointers. The keys themselves are integers, and
are used as the unique key (a suitable case within Buxton).

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>