Add BuxtonHashmap
BuxtonHashmap is a simple key-based hashmap with a focus on speed
and performance. It uses BuxtonList to back the bucket store,
using prepend operations to maintain the buckets.

This implementation uses a trivial hash and compare functions,
and may be expanded in the future. The main focus now is to
enable the replacement of the client list item list, by using
FD based lookups of clients, introducing performance gains.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>