demo: Add GUI client demo
This demo program is used to expose any potential issues within the
client API, and to demonstrate how Buxton should be used within real
programs.

Once client-side notification support is implemented, this GTK3
demo program can be used to demonstrate and perfect the techniques,
allowing to opt out and into notifications, and observe the difference
in returned keys from the layer configuration.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>