demo/gtk_client: Implement notification support
In order to facilitate the implementation of client notifications,
we must implement them in real use cases, such as a GUI program.
Currently the demo merely sets a key, and waits for a notification
of a change.

This can be tested by running two bxt_gtk_client instances, or just
one, while updating using buxtonctl. Note that this implementation
serves to expose any pitfalls and bugs in the notification
mechanism, and may be erronous in its own implementation right now.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>