core: Convert notification lists to BuxtonList
BuxtonList provides a managed solution to quickly and easily store
clients registered for notifications in the order they registered.
The previous map implemenation proved to be buggy, and broke
client notification support.

With this commit, one can observe that they can open as many
bxt_gtk_client instances as they can, and notifications will be sent
to all of the clients instantly.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>