libbuxton: Enable synchronous style get calls
If buxton_client_get_value* methods are used with the sync parameter
enabled, a client still has to provide a callback to access the data.

With this commit, a client application can set a NULL callback, enable
sync and point "data" to an initialized BuxtonData pointer, and have
the data copied into the struct should the operation prove successful.

This enables clients to immediately access the data from the call,
and manage the data within their own execution stack.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>