shared: Notify clients when CHANGED is recieved
When we recieve a CHANGED status, we use our notify callbacks map
to determine who should recieve the notification. This will require
the exposure of a poll() type method as seen in libbuxton, enabling
notifications to be checked every so often from the main loop of
a real world program.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>