direct: forbid set-label for non-root clients
The proper check here is to inspect the clients' capabilities to ensure
they have CAP_MAC_ADMIN, but this is the difficult problem to solve in a
race-free way (PID exhaustion attack, etc.).

For now, we will check the client's UID and disallow set-label for
non-root clients. This does not account for situations where we have
connections from root clients without CAP_MAC_ADMIN or non-root clients
with the capability, but these situations are unlikely at present.

Also, fix two tests that are setting labels to account for the new UID
restriction.