test: pass the correct userdata arg for set-label test
The list of arguments to the buxton_client_set_label() test was not
verified for accuracy, and the issue went unnoticed because 'make check'
must be run as root for this code to run.

Pass the group name for the userdata, which is the data the callback
expects to see.