demo: Implement notification support in the GTK client
In the GTK client we now make use of GLib's own poll functionality from
within the main loop, to execute the buxton_client_poll_response only
when we know data is available on the socket.

This allows truly asynchronous notifications, as well as results to
the buxton_client_get_value callback.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>