demo: Ignore get_value result when length is invalid
Invalid length invariably means we failed to retrieve the value,
however we'll first need to split some enums from the internal library
with the new upcoming API changes to make this more friendly.

For now, this approach is appropriate when the length not 3, i.e.
status and key name only.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>