shared: Seperate callback handling into own function
Because of the special cases involved it is cleaner to add, remove and execute
callbacks in response to daemon messages in a specific function.

Fix memory leak of notify_value and BuxtonData list.

Also keep iterating through the fd data when a single message has a problem
that isn't going to impact the rest of the messages.