core: Test for get_value returning data
In order to avoid terminating clients with a valid get_value request where no
key exists, check to data was found before trying to add it to the response. If
data isn't checked data fails to be added to the response (because it is null)
and the function to return a failure after which the client's connection is
terminated.