api: Completely rework the client API
A client should no longer use any non opaque buxton structures or non
primatives and should be able access or set any data needed through api
calls.

The most impactful changes here involve reworking what a buxton key is,
changing from a simple character string into a complex structure to allow
for complete encapsulation of the idea of a key. Looking up values based
on a key needed to be reworked in order to make fine grained access to a
key's elements simpler so the key portion of a key's structure changed to
two components a group and a name each encapsulated in a BuxtonString.