demo: Cleanup GTK demo a bit
This commit removes some redundancy from the demonstration, by making use
of the new API to enable only one callback for all Buxton calls.
We also add more flexibility to our buxton_callback by testing both the
name and type of the key in the callback, whether we've been executed
as the result of a get_value operation or notification.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>