test: use '*' data labels where appropriate
The real tests for Smack label enforcement will live in check_smack.c,
so make sure the rest of the tests can always read and write data if
this would normally be prohibited due to the client's Smack label.

This is done by using the '*' (star) label, granting world read/write
access, unless the client lacks the required DAC permissions.

The one exception to this rule, for now, is the "default label" of
floor, which is assigned to a key when the client label is absent.