test: change System/User rules in test.load2
Since the tests may be built and run on Tizen 3.0, which uses a
three-domain model with Smack enabled, it is possible that 'make check'
will run with any of "_", "System", or "User" labels.

For upcoming tests to be included in check_smack.c, a few new Smack
rules need to be added to account for this three-domain model and to
work with buxton's notion of a default label (currently "_").

To make sure existing tests behave sanely when run under Tizen 3.0, make
sure "System" and "User" labels have mutual read/write access to the
default label.