direct: set layer UID for set_label
By not setting the layer UID before calling set_value here, the daemon
will open the user layer for the daemon itself. Instead, the daemon
should open the user layer for that user.

In a future commit, set_label will be disallowed for non-root clients,
but for consistency, make sure the UID is set.