shared: Check group exists first before label test
In get_value_for_layer if smack isn't enabled then the label will not be
set and so a valid test for group existance will be skipped. To avoid this,
check the group exists if the key has a name otherwise continue as normal
and the smack check will only test group access if a client label is
passed.

Also fix up tests to not pass a client label if smack is not enabled on the
system.