test: fix set-label check on Smack systems
The label "handle_label" is not necessarily to recognized label on a
Smack-enabled system, but the subsequent gets and sets to the group with
this label may fail.

The solution is to use a Smack builtin label, "*", instead, which
permits read and write access for all clients.

Note that this solution only makes sense for tests that are not
explicitly depending on Smack access control being enforced; tests in
check_smack.c and external Buxton consumers will generally not use the
"*" label for either groups or keys.