Add API for create_group
Until now, set_label was serving three purposes:

- add/modify a label for a group
- add/modify a label for a key (currently, support is incomplete)
- create a group

To simplify the libbuxton API and the required access checking logic,
we've decided to extend the API with create_group, which will replace
the last action item in the above list.

The logic for set_label is updated in the following commit.