db: Fix set label needing correct type for key-name
When running set label, the key's data type is expected to be STRING but
this does not need to be the value's data type because set label does not
update the value. In order to allow this behavior, change set label to
run db set_value without running direct_get_value_for_layer. Then the db
will need to get the data value itself when it is passed a NULL for data
in order to store that value when writing the new label.