cli: Implement list_keys
We now implement list_keys solely within the CLI element of buxton.
Whilst it works here in a simple fashion (i.e. listing keys) it is
very far from the eventual API we wish to support.

Whilst its possible to call buxton_client_list_keys, the name is
kept different on purpose to discourage its use right now, and
internally it uses a BuxtonList, which is not yet currently exported
to public API consumers, ensuring it cannot yet be used outside
of buxtonctl.

Signed-off-by: Michael Ikey Doherty <michael.i.doherty@intel.com>