core: add read-only layer type
This patch introduce Access field for buxton's configuration. This field
isn't required, instead of another field, so get_ini_string was extended
for this purpose.

Special mode for BuxtonLayer is implemented as readonly field, but not
bitmask.

Now only gdbm backend is
supported and only set_value use cases is covered.

Signed-off-by: Alexey Perevalov <a.perevalov@samsung.com>