Simplifying public Buxton API
BuxtonSimple is an API that allows users who are new to Buxton or
familiar with vconf to utilize Buxton. There are fewer function
arguments for each function.  This is possibile because there are
default call back functions and a global client and fd provided in
the library. There are also newer functions similar to those found
in the vconf API (such has set_int32). The opening and closing
of the client connection is embedded into each function called so
the programmer does not have to call buxton_open() or buxton_close().

There is a library called libbuxtonsimp (-lbuxtonsimp) that can be
linked to by a client program when using BuxtonSimple. This exposes
only the functions in the BuxtonSimple API, without the functions
from the original Buxton API.

Man pages for the BuxtonSimple API and all public functions are
included as well as demo code. Spec file and other files needed to
build a BuxtonSimple RPM are included in the packaging directory.