Simplifying public Buxton API
BuxtonSimple is an API that allows users who are new to Buxton or
familiar with vconf to utilize Buxton. There are fewer function
arguments for each function.  This is possibile because there are
default call back functions and a global client and fd provided in
the library. There are also newer functions similar to those found
in the vconf API (such has set_int32). The opening of the client
connection is embedded into each function called so the programmer
does not have to call buxton_open().